[CPU] Move tutorials to python/tutorials/cpu
Move CPU-focused tutorials under python/tutorials/cpu, restore the
top-level tutorials to their upstream versions, and update README and CI
paths.

Tests and validation:

    python3 - <<'PY'
    from pathlib import Path
    for path in sorted(Path('python/tutorials/cpu').glob('*.py')):
        compile(path.read_text(encoding='utf-8'), str(path), 'exec')
        print(f'PASS {path}')
    PY

Result: all 9 CPU tutorial Python files compiled successfully without
writing bytecode.

    git diff --exit-code 72259b1cc -- \
      python/tutorials/01-vector-add.py \
      python/tutorials/04-low-memory-dropout.py \
      python/tutorials/05-layer-norm.py \
      python/tutorials/07-extern-functions.py

Result: all four restored top-level tutorials match upstream
byte-for-byte.

    python3 - <<'PY'
    from pathlib import Path
    import re

    sources = [
        Path('README.md'),
        *sorted(Path('.github').rglob('*.yml')),
        *sorted(Path('.github').rglob('*.yaml')),
    ]
    pattern = re.compile(r'python/tutorials/cpu/[A-Za-z0-9_.-]+\.py')
    refs = {}
    for source in sources:
        text = source.read_text(encoding='utf-8')
        for match in pattern.finditer(text):
            refs.setdefault(match.group(0), []).append(source)
    for ref, owners in sorted(refs.items()):
        exists = Path(ref).is_file()
        print(
            f'{"PASS" if exists else "FAIL"} {ref} <- '
            f'{", ".join(map(str, owners))}'
        )
        if not exists:
            raise SystemExit(1)
    print(f'TOTAL {len(refs)} unique CPU tutorial references resolved')
    PY

Result: all 3 unique README/CI references resolved:
python/tutorials/cpu/01-vector-add.py,
python/tutorials/cpu/04-blocked-matmul.py, and
python/tutorials/cpu/08-sfc-matmul.py.
