+++ python/tutorials/cpu/02-fused-softmax.py
