+++ python/tutorials/cpu/05-layer-norm.py
@@ -295,6 +295,9 @@ class LayerNorm(torch.autograd.Function):
 
 
 layer_norm = LayerNorm.apply
+device = triton.runtime.driver.active.get_current_target().backend
+# Torch doesn't support operations in float16 on CPU so use float32 instead
+dtype = torch.float32 if device == 'cpu' else torch.float16
 
 
 def test_layer_norm(M, N, dtype, eps=1e-5, device=DEVICE):
@@ -333,7 +336,7 @@ def test_layer_norm(M, N, dtype, eps=1e-5, device=DEVICE):
         styles=[('blue', '-'), ('green', '-'), ('orange', '-')],
         ylabel='GB/s',
         plot_name='layer-norm-backward',
-        args={'M': 4096, 'dtype': torch.float16, 'mode': 'backward'},
+        args={'M': 4096, 'dtype': dtype, 'mode': 'backward'},
     ))
 def bench_layer_norm(M, N, dtype, provider, mode='backward', eps=1e-5, device=DEVICE):
     # create data
@@ -371,8 +374,8 @@ def bench_layer_norm(M, N, dtype, provider, mode='backward', eps=1e-5, device=DE
     return gbps(ms), gbps(max_ms), gbps(min_ms)
 
 
-test_layer_norm(1151, 8192, torch.float16)
-bench_layer_norm.run(save_path='.', print_data=True)
+test_layer_norm(1151, 8192, dtype, device=device)
+bench_layer_norm.run(save_path='.', print_data=True, device=device)
 
 # %%
 # References
