+++ python/tutorials/cpu/gilbert_d2xy.py
